Changed reward type from float to double.
Former-commit-id: c52df7c7c303cbf4db6474d4a9407239f884abf6
Former-commit-id: 604956a033919b35c7b68540f824de8e2392d740
Former-commit-id: d22c37c4459d8bf00bbe5827e29219e8539d2cf0

+++ vizia_api_bindings/python/ViziaDoomGamePython.cpp
@@ -63,14 +63,14 @@ namespace Vizia {
         
     }
 
-    float DoomGamePython::makeAction(boost::python::list &action)
+    double DoomGamePython::makeAction(boost::python::list &action)
     {
         this->setAction(action);
         DoomGame::advanceAction();
         return DoomGame::getLastReward();
     }
 
-    float DoomGamePython::makeAction(boost::python::list &action, unsigned int tics)
+    double DoomGamePython::makeAction(boost::python::list &action, unsigned int tics)
     {
         this->setAction(action);
         DoomGame::advanceAction(true, true, tics);
